consul exec/check support for restricted shells
Enable consul exec to directly invoke an executable
file, without spawning a shell.  This enables consul
to run as restricted Linux users that have /bin/false
or /bin/nologin set as their shell, such as the
nobody account.

Consul can run as nobody, however, consul check scripts
will always return 1 (warning) because the shell is
disabled.

This change gives the operator the choice to invoke
consul exec (or check scripts) using a shell, or directly
invoke without a shell.

This change is backwards compatible for existing uses cases
that invoke binaries, such as the docs example of uptime:

    sh -c uptime

is generally equivalent to

    uptime

Existing uses cases invoking ad hoc scripts will need to
include the shell command:

    consul exec ls -lah \'some file\'

becomes:

    consul exec '$SHELL' -c ls -lah \'some file\'

This change is inspired by the Dockerfile CMD operator, which
supports running a command with or without a shell.